some changes in frontend

+++ frontend/src/admin/productList.jsx
@@ -2,7 +2,8 @@ import { Fragment, useEffect } from "react"
 import { Button } from "react-bootstrap"
 import { useDispatch, useSelector } from "react-redux"
 import { Link } from "react-router-dom"
-import {  getAdminProducts } from "../actions/productsActions"
+import {  getAdminProducts } from "../actions/productsActions";
+import { deleteProduct } from "../actions/productActions";
 import { clearError, clearProductDeleted } from "../slices/productSlice"
 import Loader from '../components/layouts/Loader';
 import { MDBDataTable} from 'mdbreact';
@@ -11,7 +12,7 @@ import Sidebar from "./sidebar";
 
 export default function ProductList() {
     const { products = [], loading = true, error }  = useSelector(state => state.productsState)
-   // const { isProductDeleted, error:productError }  = useSelector(state => state.productState)
+    const { isProductDeleted, error:productError }  = useSelector(state => state.productState)
     const dispatch = useDispatch();
 
     const setProducts = () => {
@@ -55,7 +56,7 @@ export default function ProductList() {
                 actions: (
                     <Fragment>
                         <Link to={`/admin/product/${product._id}`} className="btn btn-primary"> <i className="fa fa-pencil"></i></Link>
-                        <Button  className="btn btn-danger py-1 px-2 ml-2">
+                        <Button onClick={e => deleteHandler(e,product._id)} className="btn btn-danger py-1 px-2 ml-2">
                             <i className="fa fa-trash"></i>
                         </Button>
                     </Fragment>
@@ -66,13 +67,13 @@ export default function ProductList() {
         return data;
     }
 
-   {/* const deleteHandler = (e, id) => {
+    const deleteHandler = (e, id) => {
         e.target.disabled = true;
         dispatch(deleteProduct(id))
-    }*/}
+    }
 
     useEffect(() => {
-        if(error ) {
+        if(error || productError ) {
             toast(error , {
                 position: "bottom-center",
                 type: 'error',
@@ -80,17 +81,17 @@ export default function ProductList() {
             })
             return
         }
-        {/*if(isProductDeleted) {
+        if(isProductDeleted) {
             toast('Product Deleted Succesfully!',{
                 type: 'success',
                 position: "bottom-center",
                 onOpen: () => dispatch(clearProductDeleted())
             })
             return;
-        }*/}
+        }
 
         dispatch(getAdminProducts)
-    },[dispatch, error])
+    },[dispatch, error,isProductDeleted])
 
 
 return (
